api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]